Update Blind.h
getnextstate is executed before updating the level variable, which results in the wrong end state.
OFF is set instead of ON
Doesn't matter is you have one key for up and another for down. But with a toogle button you are two time in UP and again two times in down...