Fix debug lines not starting at line start.
Terminal programs commonly expect newlines to have both CR and LF, so
comply to that. Without doing that, those terminal programs (e.g.
ckermit) start the new line at a column that matches the line length of
the previous line - IOW, they're following the meaning of 'line feed' to
the letter.
This also matches what the Arduino classes are doing.