"Simuliere Tastendruck" implemented
The switch message is send from the central to the actor to simulate a peer message, this could be for an internal key, but also for a peer connection. The actor has to answer on the switch message with an ACK but process the message similar to an Event message.