led off for motion sensor
there is a "LED-Leuchtzeit (gn/rt)" setting   which sets the led on duration if a motion was recognised.
The value can be 0 as well, which shall switch off the led, but the led was still blinking for the transition.
My changes ensure on a 0 value that the led is switched off…