Add files via upload
Extension of the encoder to half and quarter steps for encoders that have 2 or 4 detent positions per pulse period.
The configuration takes place when the class instance is created using the constructor. The new parameters are optional, so existing code can continue to be used unchanged.

A special feature emerged with double or quadruple evaluation when using hardware debouncing of the encoder signals via RC element: If the interrupt release occurs during initialization immediately after the port mode has been defined, the capacitor is still empty and is charging. This leads to an unwanted detection of an alleged rotary movement. Therefore the interrupt release must be delayed. For the sake of simplicity, I used the delay() method here, since this is called only once during boot and only during setup.